+++ CHANGELOG.md
@@ -5,10 +5,14 @@
 
 ### added
 - added `asdf.util.get_schema_path` helper function [[#325]](https://github.com/BAMWelDX/weldx/pull/325)
+- added `util.compare_nested` to check equality of two nested data structures. [[#328]](https://github.com/BAMWelDX/weldx/pull/328)
 
 ### changes
 - `WXRotation.from_euler()` now accepts a `pint.Quantity` as input. [[#318]](https://github.com/BAMWelDX/weldx/pull/318)
 - move tests folder to `weldx/tests` [[#323]](https://github.com/BAMWelDX/weldx/pull/323)
+- `get_yaml_header` received a new option parse, which optionally returns the parsed YAML
+  header as `asdf.tagged.TaggedDict`. [[#338]](https://github.com/BAMWelDX/weldx/pull/338)
+
 
 ## 0.3.3 (30.03.2021)